Remove the RSDConditionalRule
This is a weird and obfuscated holdover from implementing a work-around to ResearchKit not allowing the `ORKTask` to mutate the task result. It isn’t used and it would be better to remove it *now* before we start releasing this framework to a broader audience.
